#33 Added Laravel 12 support.
