+++ composer.json
@@ -38,12 +38,12 @@
 		}
 	},
 	"require": {
-		"illuminate/console" : "^7.0|^8.0|^9.0|^10.0|^11.0",
-		"illuminate/support" : "^7.0|^8.0|^9.0|^10.0|^11.0",
-		"illuminate/database": "^7.0|^8.0|^9.0|^10.0|^11.0"
+		"illuminate/console": "^7.0|^8.0|^9.0|^10.0|^11.0|^12.0",
+		"illuminate/support": "^7.0|^8.0|^9.0|^10.0|^11.0|^12.0",
+		"illuminate/database": "^7.0|^8.0|^9.0|^10.0|^11.0|^12.0"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "^7.0|^8.0|^9.0|^10.0|^11.0",
+		"phpunit/phpunit": "^7.0|^8.0|^9.0|^10.0|^11.0|^12.0",
 		"orchestra/testbench": "^7.0|^8.0|^9.0|^10.0|^11.0"
 	}
 }
