+++ src/Console/Commands/BuilderMakeCommand.php
@@ -2,7 +2,6 @@
 
 namespace Bjnstnkvc\BuilderMakeCommand\Console\Commands;
 
-use Bjnstnkvc\BuilderMakeCommand\Clauses\{OrWhere, OrWhereIn, OrWhereLike, OrWhereNot, OrWhereNotIn, OrWhereNotLike, Where, WhereIn, WhereLike, WhereNot, WhereNotIn, WhereNotLike};
 use Illuminate\Console\GeneratorCommand;
 use Illuminate\Contracts\Console\PromptsForMissingInput;
 use Illuminate\Contracts\Filesystem\FileNotFoundException;
@@ -70,15 +69,15 @@ class BuilderMakeCommand extends GeneratorCommand implements PromptsForMissingIn
 
         $this->newLine();
 
-        $this->info(
-            "/**\n* Create a new Eloquent query builder for the model\n*\n* @param \$query\n*\n* @return UserBuilder\n*/\npublic function newEloquentBuilder(\$query): UserBuilder\n{\n\treturn new UserBuilder(\$query);\n}"
-        );
+        $this->info("/**\n* Create a new Eloquent query builder for the model\n*\n* @param \$query\n*\n* @return UserBuilder\n*/\npublic function newEloquentBuilder(\$query): UserBuilder\n{\n\treturn new UserBuilder(\$query);\n}");
 
         $this->newLine();
 
-        $this->info(
-            "/**\n* @method static UserBuilder query() Begin querying the model.\n*\n* @mixin {$this->argument('name')}\n*/"
-        );
+        $this->comment("You can also add the following doc comment to your {$this->argument('model')} model to enhance your editors intellisense:");
+
+        $this->newLine();
+
+        $this->info("/**\n* @method static UserBuilder query() Begin querying the model.\n*\n* @mixin {$this->argument('name')}\n*/");
 
         $this->newLine();
 
@@ -230,20 +229,15 @@ class BuilderMakeCommand extends GeneratorCommand implements PromptsForMissingIn
      */
     protected function setMethodSignaturesForColumn(string $column): array
     {
-        return [
-            Where::make($column)->signature(),
-            WhereNot::make($column)->signature(),
-            WhereIn::make($column)->signature(),
-            WhereNotIn::make($column)->signature(),
-            WhereLike::make($column)->signature(),
-            WhereNotLike::make($column)->signature(),
-            OrWhere::make($column)->signature(),
-            OrWhereNot::make($column)->signature(),
-            OrWhereIn::make($column)->signature(),
-            OrWhereNotIn::make($column)->signature(),
-            OrWhereLike::make($column)->signature(),
-            OrWhereNotLike::make($column)->signature(),
-        ];
+        return collect(glob(__DIR__ . '/../../Clauses/*.php'))
+            ->map(function (string $path) use ($column) {
+                return Str::of($path)
+                    ->basename('.php')
+                    ->prepend('Bjnstnkvc\BuilderMakeCommand\Clauses\\')
+                    ->pipe(fn(string $clause) => $clause::make($column)->signature())
+                    ->toString();
+            })
+            ->toArray();
     }
 
     /**
