--- workbench/app/Clauses/InvalidClause.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace App\Clauses;
-
-use Bjnstnkvc\BuilderMakeCommand\Abstracts\Clause;
-
-/**
- * @method static static make(string $column) Make a new clause instance.
- */
-class InvalidClause extends Clause
-{
-    /**
-     * Determine if the clause is the given method.
-     *
-     * @param string $method
-     *
-     * @return bool
-     */
-    public static function is(string $method): bool
-    {
-        return $method === 'invalid';
-    }
-}
