--- workbench/database/factories/TestUserFactory.php
@@ -1,39 +1,0 @@
-<?php
-
-namespace Database\Factories;
-
-use App\Models\TestUser;
-use Illuminate\Database\Eloquent\Factories\Factory;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Str;
-
-class TestUserFactory extends Factory
-{
-    /**
-     * The name of the factory's corresponding model.
-     *
-     * @var class-string
-     */
-    protected $model = TestUser::class;
-
-    /**
-     * The current password being used by the factory.
-     */
-    protected static ?string $password;
-
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
-    public function definition(): array
-    {
-        return [
-            'name'              => fake()->name(),
-            'email'             => fake()->unique()->safeEmail(),
-            'email_verified_at' => now(),
-            'password'          => static::$password ??= Hash::make('password'),
-            'remember_token'    => Str::random(10),
-        ];
-    }
-}
